+++ INTERX/gateway/main.go
@@ -2,6 +2,7 @@ package gateway
 
 import (
 	"context"
+	"crypto/tls"
 	"fmt"
 	"mime"
 	"net/http"
@@ -9,11 +10,13 @@ import (
 	"github.com/KiraCore/sekai/INTERX/config"
 	"github.com/KiraCore/sekai/INTERX/database"
 	"github.com/KiraCore/sekai/INTERX/functions"
+	"github.com/KiraCore/sekai/INTERX/insecure"
 	cosmosAuth "github.com/KiraCore/sekai/INTERX/proto-gen/cosmos/auth"
 	cosmosBank "github.com/KiraCore/sekai/INTERX/proto-gen/cosmos/bank"
 	kiraGov "github.com/KiraCore/sekai/INTERX/proto-gen/kira/gov"
 	kiraSlashing "github.com/KiraCore/sekai/INTERX/proto-gen/kira/slashing"
 	kiraStaking "github.com/KiraCore/sekai/INTERX/proto-gen/kira/staking"
+	kiraTokens "github.com/KiraCore/sekai/INTERX/proto-gen/kira/tokens"
 	"github.com/KiraCore/sekai/INTERX/tasks"
 	functionmeta "github.com/KiraCore/sekai/function_meta"
 	"github.com/gorilla/mux"
@@ -86,6 +89,11 @@ func GetGrpcServeMux(grpcAddr string) (*runtime.ServeMux, error) {
 		return nil, fmt.Errorf("failed to register gateway: %w", err)
 	}
 
+	err = kiraTokens.RegisterQueryHandler(context.Background(), gwCosmosmux, conn)
+	if err != nil {
+		return nil, fmt.Errorf("failed to register gateway: %w", err)
+	}
+
 	return gwCosmosmux, nil
 }
 
@@ -99,6 +107,7 @@ func Run(configFilePath string, log grpclog.LoggerV2) error {
 	database.LoadFaucetDbDriver()
 	database.LoadReferenceDbDriver()
 
+	serveHTTPS := config.Config.ServeHTTPS
 	grpcAddr := config.Config.GRPC
 	rpcAddr := config.Config.RPC
 	port := config.Config.PORT
@@ -128,15 +137,14 @@ func Run(configFilePath string, log grpclog.LoggerV2) error {
 		Handler: c.Handler(router),
 	}
 
-	// SERVE_HTTP: Empty parameters mean use the TLS Config specified with the server.
-	// if strings.ToLower(os.Getenv("SERVE_HTTP")) == "false" {
-	// 	gwServer.TLSConfig = &tls.Config{
-	// 		Certificates: []tls.Certificate{insecure.Cert},
-	// 	}
+	if serveHTTPS {
+		gwServer.TLSConfig = &tls.Config{
+			Certificates: []tls.Certificate{insecure.Cert},
+		}
 
-	// 	log.Info("Serving gRPC-Gateway and OpenAPI Documentation on https://", gatewayAddr)
-	// 	return fmt.Errorf("serving gRPC-Gateway server: %w", gwServer.ListenAndServeTLS("", ""))
-	// }
+		log.Info("Serving gRPC-Gateway and OpenAPI Documentation on https://", gatewayAddr)
+		return fmt.Errorf("serving gRPC-Gateway server: %w", gwServer.ListenAndServeTLS("", ""))
+	}
 
 	log.Info("Serving gRPC-Gateway and OpenAPI Documentation on http://", gatewayAddr)
 	return fmt.Errorf("serving gRPC-Gateway server: %w", gwServer.ListenAndServe())