+++ x/gov/keeper/grpc_query.go
@@ -109,7 +109,6 @@ func (q Querier) GetPoorNetworkMessages(ctx context.Context, request *types.Poor
 
 // Proposal returns a proposal by id
 func (q Querier) Proposal(ctx context.Context, request *types.QueryProposalRequest) (*types.QueryProposalResponse, error) {
-	fmt.Println("Proposal request", request)
 	sdkContext := sdk.UnwrapSDKContext(ctx)
 	proposal, found := q.keeper.GetProposal(sdkContext, request.ProposalId)
 	if found == false {
@@ -118,16 +117,14 @@ func (q Querier) Proposal(ctx context.Context, request *types.QueryProposalReque
 	votes := q.keeper.GetProposalVotes(sdkContext, request.ProposalId)
 	return &types.QueryProposalResponse{
 		Proposal: proposal,
-		Votes: votes,
+		Votes:    votes,
 	}, nil
 }
 
 // Proposals query proposals by querying params
 func (q Querier) Proposals(ctx context.Context, request *types.QueryProposalsRequest) (*types.QueryProposalsResponse, error) {
-	fmt.Println("proposals request", request)
 	sdkContext := sdk.UnwrapSDKContext(ctx)
 	proposals, err := q.keeper.GetProposals(sdkContext)
-	fmt.Println("proposals", proposals, err)
 	if err != nil {
 		return nil, sdkerrors.Wrap(types.ErrGettingProposals, fmt.Sprintf("error getting proposals: %s", err.Error()))
 	}